+++ src/main/kotlin/no/nav/syfo/common/journalforing/JournalpostId.kt
@@ -1,0 +1,6 @@
+package no.nav.syfo.common.journalforing
+
+@JvmInline
+public value class JournalpostId(
+    public val value: Int,
+)